refactor(engine): improve types

+++ game-engine/src/ai/mod.rs
@@ -4,6 +4,6 @@ use crate::{BoardDefinition, Ship, Pos, BoardState};
 
 pub trait Engine {
     fn get_name(&self) -> String;
-    fn place_ships(&mut self, board: &BoardDefinition, ships: Vec<i32>) -> Vec<Ship>;
+    fn place_ships(&mut self, board: &BoardDefinition, ships: Vec<usize>) -> Vec<Ship>;
     fn get_shot(&mut self, board: &BoardState) -> Pos;
 }
